fix build

+++ src/test.rs
@@ -9,6 +9,6 @@ static mut REPLICA_COUNT: AtomicUint = INIT_ATOMIC_UINT;
 /// The replica ID is guaranteed to be unique within the processes. This
 /// function should **not** be used for generating replica IDs in a distributed
 /// system.
-pub fn gen_replica_id() -> uint {
-    unsafe { REPLICA_COUNT.fetch_add(1, SeqCst) }
+pub fn gen_replica_id() -> u64 {
+    unsafe { REPLICA_COUNT.fetch_add(1, SeqCst) as u64 }
 }
